tests_python/requirements.txt
pytest
pyserial
pytest-timeout
pytest-html
